fix(view, html): move to js file
