+++ index.html
@@ -285,7 +285,7 @@
                 </a>
               </li>
               <li class="list-item__header-menu">
-                <a href="./src/page/product.html">
+                <a href="./src/page/product/product.html">
                   <span class="text__menu-title">Products</span>
                 </a>
               </li>
@@ -1469,7 +1469,7 @@
           </strong>
           <div class="box__gallery">
             <div class="box__gallery-item section__gallery-large">
-              <a href="./src/page/blog.html">
+              <a href="./src/page/blog/blog.html">
                 <div class="box__gallery-image">
                   <div class="box__image">
                     <img
@@ -1488,7 +1488,7 @@
             </div>
             <div class="section__gallery-large">
               <div class="box__gallery-item">
-                <a href="./src/page/blog.html">
+                <a href="./src/page/blog/blog.html">
                   <div class="box__gallery-image">
                     <div class="box__image">
                       <img
@@ -1507,7 +1507,7 @@
               </div>
               <div class="section__gallery-bottom">
                 <div class="box__gallery-item section__gallery-small">
-                  <a href="./src/page/blog.html">
+                  <a href="./src/page/blog/blog.html">
                     <div class="box__gallery-image">
                       <div class="box__image">
                         <img
@@ -1525,7 +1525,7 @@
                   </a>
                 </div>
                 <div class="box__gallery-item section__gallery-small">
-                  <a href="./src/page/blog.html">
+                  <a href="./src/page/blog/blog.html">
                     <div class="box__gallery-image">
                       <div class="box__image">
                         <img
