+++ src/page/home/home.js
@@ -1,14 +1,13 @@
 // TODO 무조건 file은 알파벳 순서대로 되어야함. 이렇게 안하면 나중에 rebase, merge 에서 다 깨짐. 불필요한 작업이 됨.
 import '../asset/style/component/footer.scss';
 import '../asset/style/component/header.scss';
-import '../asset/style/media/blogMedia.scss';
 import '../asset/style/media/footerMedia.scss';
 import '../asset/style/media/headerMedia.scss';
-import '../asset/style/page/blog.scss';
+import '../asset/style/media/homeMedia.scss';
+import '../asset/style/page/home.scss';
 
-import '../component/bannerSlider';
-import '../component/blog/scrollIndicator';
-import '../component/blog/scrollMenu';
-import '../component/blog/scrollTopBlock';
-import '../component/homeAccordion';
-import '../component/homeSideBar';
+import '../../component/bannerSlider';
+import '../../component/homeAccordion';
+import '../../component/homeSearch';
+import '../../component/homeSideBar';
+import '../../component/MainSlider';
