+++ src/page/productList/products.js
@@ -3,15 +3,15 @@ import "../asset/style/component/header.scss";
 import "../asset/style/media/headerMedia.scss";
 import "../asset/style/page/products.scss";
 
-import "../component/bannerSlider";
-import "../component/cart/toggle";
-import "../component/cart/update";
+import "../../component/bannerSlider";
+import "../../component/cart/toggle";
+import "../../component/cart/update";
 
 import { getProducts } from "../api/product";
-import { storeState, setStoreState } from "../store/store";
+import { storeState, setStoreState } from "../../store/store";
 import { addItemsToContainer } from "../util/render";
-import renderSearchComponent from "../component/search";
-import { renderCompanyComponent } from "../component/companies";
+import renderSearchComponent from "../../component/search";
+import { renderCompanyComponent } from "../../component/companies";
 
 const initializeApp = async () => {
   const productContainer = document.querySelector(".list__card-products");
